+++ CHANGELOG.rst
@@ -4,7 +4,10 @@ CHANGELOG
 7.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Feat: Add metadata info to workflows
+- Fix: Update workflow vocabulary name
+- Feat: Update workflow vocabulary title attribute to use metadata
+  [rboixaderg]
 
 
 7.0.0 (2023-12-06)