+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ CHANGELOG
 7.0.1 (unreleased)
 ------------------
 
+- Being able to use schema.Time
+[nilbacardit26]
 - Feat: Add metadata info to workflows
 - Fix: Update workflow vocabulary name
 - Feat: Update workflow vocabulary title attribute to use metadata